pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "short-links-service"
version = "0.1.0"
description = "Scaffold for a short links service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.16.0,<2.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "pydantic-settings>=2.8.0,<3.0.0",
  "redis>=6.0.0,<7.0.0",
  "sqlalchemy>=2.0.0,<3.0.0",
  "streamlit>=1.44.0,<2.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
